Meat and Potatoes
Made some minor edits along with some helper functions. The major edit: I created the turn-based aspect of the game, such as the teams, who's turn it is, the AI framework, etc. As of now, I'm testing out the combat system with a mock battle. Some issues are arising, but I'm debugging them as they come. I like how far its come so far and definitely plan on going further.