Items_pt2
Still working on the items, but I got the usage of them down. Going to make the storage a library to clean things up and make it faster and easier to access. The player can use items. The main problems are updating the storage when an item is used and the base case of what to do when the storage is empty.