Moving forward
So, I think I'm done with this project, that is to say I've lost motivation in completing it and making it a fully functional game. I want to start on other projects and  this project has helped me tremendously in seeing what all goes into an rpg. I want to state what few things I intend to implement before moving on. #1, a way to save/load game states from a file. #2, a way to auto generate enemies. #3, a survival mode, and #4, a little story mode. Hopefully I can have all that implemented by the end of the year.